backend/requirements.txt
fastapi
uvicorn
requests
pandas
python-dotenv
openai
twilio
pydantic
